+++ src/include/LSP/LuauExt.hpp
@@ -119,13 +119,19 @@ lsp::Diagnostic createParseErrorDiagnostic(const Luau::ParseError& error, const 
 bool isGetService(const Luau::AstExpr* expr);
 bool isRequire(const Luau::AstExpr* expr);
 
-struct FindServicesVisitor : public Luau::AstVisitor
+struct FindImportsVisitor : public Luau::AstVisitor
 {
+private:
+    std::optional<size_t> previousRequireLine = std::nullopt;
+
+public:
     std::optional<size_t> firstServiceDefinitionLine = std::nullopt;
     std::optional<size_t> lastServiceDefinitionLine = std::nullopt;
     std::map<std::string, Luau::AstStatLocal*> serviceLineMap{};
+    std::optional<size_t> firstRequireLine = std::nullopt;
+    std::vector<std::map<std::string, Luau::AstStatLocal*>> requiresMap{{}};
 
-    size_t findBestLine(const std::string& serviceName, size_t minimumLineNumber)
+    size_t findBestLineForService(const std::string& serviceName, size_t minimumLineNumber)
     {
         if (firstServiceDefinitionLine)
             minimumLineNumber = *firstServiceDefinitionLine > minimumLineNumber ? *firstServiceDefinitionLine : minimumLineNumber;
@@ -138,6 +144,14 @@ struct FindServicesVisitor : public Luau::AstVisitor
                 lineNumber = location + 1;
         }
         return lineNumber;
+    }
+
+    bool containsRequire(const std::string& module)
+    {
+        for (const auto& map : requiresMap)
+            if (contains(map, module))
+                return true;
+        return false;
     }
 
     bool visit(Luau::AstStatLocal* local) override
@@ -161,54 +175,9 @@ struct FindServicesVisitor : public Luau::AstVisitor
                 !lastServiceDefinitionLine.has_value() || lastServiceDefinitionLine.value() <= line ? line : lastServiceDefinitionLine.value();
             serviceLineMap.emplace(std::string(localName->name.value), local);
         }
-
-        return false;
-    }
-
-    bool visit(Luau::AstStatBlock* block) override
-    {
-        for (Luau::AstStat* stat : block->body)
-        {
-            stat->visit(this);
-        }
-
-        return false;
-    }
-};
-
-struct FindRequiresVisitor : public Luau::AstVisitor
-{
-    std::optional<size_t> firstRequireLine = std::nullopt;
-    std::optional<size_t> previousRequireLine = std::nullopt;
-    std::vector<std::map<std::string, Luau::AstStatLocal*>> requiresMap{{}};
-
-    bool contains(const std::string& module)
-    {
-        for (auto& map : requiresMap)
-        {
-            if (map.find(module) != map.end())
-                return true;
-        }
-        return false;
-    }
-
-    bool visit(Luau::AstStatLocal* local) override
-    {
-        if (local->vars.size != 1 || local->values.size != 1)
-            return false;
-
-        auto localName = local->vars.data[0];
-        auto expr = local->values.data[0];
-
-        if (!localName || !expr)
-            return false;
-
-        auto line = localName->location.begin.line;
-
-        if (isRequire(expr))
+        else if (isRequire(expr))
         {
             firstRequireLine = !firstRequireLine.has_value() || firstRequireLine.value() >= line ? line : firstRequireLine.value();
-
 
             // If the requires are too many lines away, treat it as a new group
             if (previousRequireLine && line - previousRequireLine.value() > 1)